+++ crates/holochain/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Add query `are_all_blocked`, taking a vector of `BlockTargetId`s, to `holochain_state`.
+- Upgrade to kitsune2 0.3.0.
+- Deprecate `BlockTarget::NodeDna` and `BlockTarget::Node`. These block variants are not currently respected and will be removed with the 0.6.0 minor release.
+
 ## 0.6.0-dev.21
 
 - Validate warrants in the sys validation workflow. \#5249